Switch to using img instead of sprites
The sprites thing doesn't seem to work with this combi of requirements:

- cross-browser, including DuckDuckGo's and Android's Chrome browser
- scaling the slices from the sprite
- keeping things responsive

So I've given up and committed raw assets to git for now, all of the
cards. I did try to help the browser prefetch all the images, so that
the app keeps working when you loose internet for a moment.